[BPK-2781]: Renamed the BpkImageCornerStyles file to BpkImageBorderRadiusStyles
