+++ packages/bpk-component-image/index.js
@@ -20,7 +20,7 @@ import BpkImage from './src/BpkImage';
 import BpkBackgroundImage from './src/BpkBackgroundImage';
 import withLazyLoading from './src/withLazyLoading';
 import withLoadingBehavior from './src/withLoadingBehavior';
-import BORDER_RADIUS_STYLES from './src/BpkImageCornerStyles';
+import BORDER_RADIUS_STYLES from './src/BpkImageBorderRadiusStyles';
 
 export default BpkImage;
